chore(react): rename Flag to CountryFlag
