+++ packages/react/src/components/CountryFlag/CountryFlag.stories.mdx
@@ -1,18 +1,18 @@
 import {ArgsTable, Source, Story, Canvas, Meta} from '@storybook/addon-docs';
-import Flag from './Flag';
+import CountryFlag from './CountryFlag';
 import dedent from 'ts-dedent';
 import StoryConfig from '../../../.storybook/story-config.ts';
 
 export const meta = {
-  component: Flag,
-  title: StoryConfig.Flag.hierarchy
+  component: CountryFlag,
+  title: StoryConfig.CountryFlag.hierarchy
 };
 
 <Meta title={meta.title} component={meta.component} />
 
-export const Template = args => <Flag {...args} />;
+export const Template = args => <CountryFlag {...args} />;
 
-# Flag
+# CountryFlag
 
 - [Overview](#overview)
 - [Props](#props)
@@ -20,7 +20,7 @@ export const Template = args => <Flag {...args} />;
 
 ## Overview
 
-Flag can be used to show the flag of the country for the provided country code.
+CountryFlag can be used to show the flag of the country for the provided country code.
 
 <Canvas>
   <Story name="Overview" args={{
@@ -36,17 +36,17 @@ Flag can be used to show the flag of the country for the provided country code.
 
 ## Usage
 
-Import and use the `Flag` component in your components as follows.
+Import and use the `CountryFlag` component in your components as follows.
 
 <Source
   language="jsx"
   dark
   format
   code={dedent`
-import Flag from '@oxygen-ui/react/Flag';\n
+import CountryFlag from '@oxygen-ui/react/CountryFlag';\n
 function Demo() {
   return (
-    <Flag
+    <CountryFlag
       title="Sri Lanka"
       countryCode="LK"
     />
